+++ src/main/java/com/practice/management/bean/entity/Task.java
@@ -3,7 +3,7 @@ package com.practice.management.bean.entity;
 import java.util.Date;
 
 public class Task {
-    private int id;
+    private Long id;
     private String projectName;
     private String year;
     private Date submitTime;
@@ -15,15 +15,15 @@ public class Task {
     private String enclosureUrl;
     private String enterpriseOpinion;
     private String schoolEamine;
-    private int enterpriseId;
-    private int schoolId;
-    private int studentId;
+    private Long enterpriseId;
+    private Long schoolId;
+    private Long studentId;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -115,27 +115,27 @@ public class Task {
         this.schoolEamine = schoolEamine;
     }
 
-    public int getEnterpriseId() {
+    public Long getEnterpriseId() {
         return enterpriseId;
     }
 
-    public void setEnterpriseId(int enterpriseId) {
+    public void setEnterpriseId(Long enterpriseId) {
         this.enterpriseId = enterpriseId;
     }
 
-    public int getSchoolId() {
+    public Long getSchoolId() {
         return schoolId;
     }
 
-    public void setSchoolId(int schoolId) {
+    public void setSchoolId(Long schoolId) {
         this.schoolId = schoolId;
     }
 
-    public int getStudentId() {
+    public Long getStudentId() {
         return studentId;
     }
 
-    public void setStudentId(int studentId) {
+    public void setStudentId(Long studentId) {
         this.studentId = studentId;
     }
 }
